requirements.txt
asgiref==3.8.1
Django==5.1.1
django-autoslug==1.9.9
django-extensions==3.2.3
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.2
